Fix issues saving options for non-"direct" filesystem type.
In the case that the filesystem type is detected as non-"direct," the
options from the lightbox when first installing the plugin are
attempted to be saved on the next page load. However, some of the
conditional checks here had incorrect $_POST variable name checks, as
well as the check_admin_referer string. So if you were in this
situation, you get the lightbox popping up repeatedly on every page
load on the admin, until you decide to fix the issue and send a pull
request.