Allow defining recommended constants by project type
Previously, constants were either recommended or not. This adds the
ability to define which project types should receive the constant
recommendation in the same way projects are defined.

This will be useful when we add Jetpack support, as Jetpack has a dev
mode constant that shouldn't be displayed for certain project types.

See #39

Fixes #42